requirements.txt
alembic==1.16.1
altgraph==0.17.4
asttokens==3.0.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
docutils==0.21.2
executing==2.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
greenlet==3.0.3
idna==3.7
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pygame==2.6.0
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.10.1
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==72.1.0
six==1.16.0
SQLAlchemy==2.0.32
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
